moved some aws_events out of test_events into test_aws_events
