+++ src/publisher/tests/test_aws_events.py
@@ -1,7 +1,7 @@
 import json
 from unittest.mock import patch, Mock
 from publisher import ajson_ingestor
-from .base import BaseCase
+from . import base
 from os.path import join
 from django.core.urlresolvers import reverse
 from django.contrib.auth.models import User
@@ -59,7 +59,7 @@ def formsubgen(art):
         })
     return artf
 
-class One(BaseCase):
+class One(base.BaseCase):
     def setUp(self):
         self.ac = Client()
 
@@ -128,3 +128,56 @@ class One(BaseCase):
             resp = self.ac.post(url, payload, follow=False)
             self.assertEqual(resp.status_code, 302) # temp redirect after successful submission
             mock.publish.assert_called_once_with(Message=expected_event)
+
+
+class Two(base.TransactionBaseCase):
+    def setUp(self):
+        self.f1 = join(self.fixture_dir, 'ajson', 'elife-10627-v1.xml.json')
+        self.ajson1 = self.load_ajson(self.f1, strip_relations=False)
+
+        self.f2 = join(self.fixture_dir, 'ajson', 'elife-09560-v1.xml.json')
+        self.ajson2 = self.load_ajson(self.f2, strip_relations=False)
+
+    def tearDown(self):
+        pass
+
+    @patch('publisher.ajson_ingestor.aws_events.notify')
+    def test_related_events(self, notify_mock):
+        "aws_events.notify is called once for the article being ingested and once each for related articles"
+
+        ajson_ingestor.ingest(self.ajson1) # has 2 related
+
+        def event_msid(index):
+            args, first_arg = 1, 0
+            return notify_mock.mock_calls[index][args][first_arg].manuscript_id
+
+        # 10627 has two relations, 9561 and 9560
+        # ensure `notify` called once for each article
+        self.assertEqual(len(notify_mock.mock_calls), 3)
+
+        # ensure the events have the right manuscript id
+        # internal relationships, lowest to highest msid
+        self.assertEqual(event_msid(1), 9560)
+        self.assertEqual(event_msid(2), 9561)
+
+    def test_related_events2(self):
+        """aws_events.notify is called once for the article being ingested and once
+        each for related articles, including reverse relations"""
+
+        ajson_ingestor.ingest(self.ajson1) # has 2 related, 9561 and 9560
+
+        with patch('publisher.ajson_ingestor.aws_events.notify') as notify_mock:
+
+            def event_msid(index):
+                args, first_arg = 1, 0
+                return notify_mock.mock_calls[index][args][first_arg].manuscript_id
+
+            ajson_ingestor.ingest(self.ajson2) # has 2 related, 10627, 9561
+
+            self.assertEqual(len(notify_mock.mock_calls), 3)
+
+            # ensure the events have the right manuscript id
+            self.assertEqual(event_msid(0), 9560)
+            # internal relationships, lowest to highest msid
+            self.assertEqual(event_msid(1), 9561) # linked by 9560
+            self.assertEqual(event_msid(2), 10627) # links to 9560
